refactor: 불필요한 접근 제어자 public 수정
